refactor: remove barrel files per AGENTS.md's no-barrel-files rule
Part of #206's hygiene sweep.

- ContextMenu/index.ts and TimePicker/index.ts were real re-export
  barrels (one consumer each). Deleted; TitleBar.tsx and EditAlarm.tsx
  now import directly from ContextMenu/ContextMenu.tsx +
  ContextMenu/types.ts and TimePicker/TimePicker.tsx respectively.
- Icons/index.tsx wasn't actually a re-export barrel -- it defined the
  icon components directly -- but TitleBar.tsx still imported it via
  directory-shorthand (`from './Icons'`), the exact pattern the rule
  bans. Renamed to Icons/Icons.tsx and updated the import.

Confirmed no other consumers of any of the three via a repo-wide grep.
