--- apps/threshold/src/components/ContextMenu/index.ts
@@ -1,2 +1,0 @@
-export { ContextMenu } from './ContextMenu';
-export type { MenuModel, MenuPosition, MenuItem, MenuSection, MenuSeparator } from './types';
