--- apps/threshold/src/components/TimePicker/index.ts
@@ -1,1 +1,0 @@
-export * from './TimePicker';
